Extract PhongLighting lib
