+++ DX12Renderer/src/LightingDemo/PointLightShadowPassPso.cpp
@@ -2,7 +2,7 @@
 #include <LightingDemo/ShadowPassPsoBase.h>
 
 #include <DX12Library/CommandList.h>
-#include <Light.h>
+#include <Framework/Light.h>
 #include <DX12Library/Cubemap.h>
 
 using namespace Microsoft::WRL;
