+++ DX12Renderer/src/LightingDemo/SceneRenderer.cpp
@@ -5,7 +5,7 @@
 #include <Framework/Model.h>
 #include <Framework/Mesh.h>
 #include <Framework/MatricesCb.h>
-#include <Material.h>
+#include <PhongLighting/Material.h>
 #include <DX12Library/Cubemap.h>
 #include <DX12Library/ShaderUtils.h>
